+++ processor/src/test/fixtures/input/com/example/component_dependency/BaseParameterizedFieldDependencyInParentModel.java
@@ -1,0 +1,11 @@
+package com.example.component_dependency;
+
+import arez.annotations.ActAsComponent;
+import arez.annotations.ComponentDependency;
+
+@ActAsComponent
+abstract class BaseParameterizedFieldDependencyInParentModel<T>
+{
+  @ComponentDependency
+  final T value = null;
+}