+++ processor/src/test/fixtures/input/com/example/component_dependency/ParameterizedFieldDependencyInParentModel.java
@@ -1,0 +1,11 @@
+package com.example.component_dependency;
+
+import arez.annotations.ArezComponent;
+import arez.component.DisposeNotifier;
+
+@ArezComponent
+abstract class ParameterizedFieldDependencyInParentModel
+  extends BaseParameterizedFieldDependencyInParentModel<DisposeNotifier>
+{
+}
+