+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Allow the type of a field annotated by `@ComponentDependency` to be parameterized in a base class and resolved in the arez component.
+
 ### [v0.179](https://github.com/arez/arez/tree/v0.179) (2020-05-21) · [Full Changelog](https://github.com/arez/arez/compare/v0.178...v0.179)
 
 Changes in this release: